+++ asyncgit/src/error.rs
@@ -1,4 +1,4 @@
-use std::string::FromUtf8Error;
+use std::{num::TryFromIntError, string::FromUtf8Error};
 use thiserror::Error;
 
 #[derive(Error, Debug)]
@@ -26,6 +26,9 @@ pub enum Error {
 
     #[error("utf8 error:{0}")]
     Utf8Error(#[from] FromUtf8Error),
+
+    #[error("TryFromInt error:{0}")]
+    IntError(#[from] TryFromIntError),
 }
 
 pub type Result<T> = std::result::Result<T, Error>;