+++ asyncgit/src/lib.rs
@@ -3,6 +3,7 @@
 #![forbid(missing_docs)]
 #![deny(unsafe_code)]
 #![deny(unused_imports)]
+#![deny(unused_must_use)]
 #![deny(clippy::all)]
 #![deny(clippy::unwrap_used)]
 #![deny(clippy::panic)]
@@ -15,7 +16,9 @@ mod commit_files;
 mod diff;
 mod error;
 mod fetch;
+mod progress;
 mod push;
+mod push_tags;
 pub mod remote_progress;
 mod revlog;
 mod status;
@@ -27,6 +30,7 @@ pub use crate::{
     diff::{AsyncDiff, DiffParams, DiffType},
     fetch::{AsyncFetch, FetchRequest},
     push::{AsyncPush, PushRequest},
+    push_tags::{AsyncPushTags, PushTagsRequest},
     remote_progress::{RemoteProgress, RemoteProgressState},
     revlog::{AsyncLog, FetchStatus},
     status::{AsyncStatus, StatusParams},
@@ -58,6 +62,8 @@ pub enum AsyncNotification {
     Tags,
     ///
     Push,
+    ///
+    PushTags,
     ///
     Fetch,
 }