+++ src/queue.rs
@@ -1,5 +1,5 @@
 use crate::tabs::StashingOptions;
-use asyncgit::sync::{CommitId, CommitTags};
+use asyncgit::sync::{diff::DiffLinePosition, CommitId, CommitTags};
 use bitflags::bitflags;
 use std::{cell::RefCell, collections::VecDeque, rc::Rc};
 
@@ -27,6 +27,7 @@ pub struct ResetItem {
 pub enum Action {
     Reset(ResetItem),
     ResetHunk(String, u64),
+    ResetLines(String, Vec<DiffLinePosition>),
     StashDrop(CommitId),
     DeleteBranch(String),
     ForcePush(String, bool),
@@ -65,6 +66,8 @@ pub enum InternalEvent {
     Push(String, bool),
     ///
     Pull(String),
+    ///
+    PushTags,
 }
 
 ///