+++ src/strings.rs
@@ -14,22 +14,21 @@ pub static PUSH_POPUP_STATES_PUSHING: &str = "pushing (3/3)";
 pub static PUSH_POPUP_STATES_TRANSFER: &str = "transfer";
 pub static PUSH_POPUP_STATES_DONE: &str = "done";
 
+pub static PUSH_TAGS_POPUP_MSG: &str = "Push Tags";
+pub static PUSH_TAGS_STATES_FETCHING: &str = "fetching";
+pub static PUSH_TAGS_STATES_PUSHING: &str = "pushing";
+pub static PUSH_TAGS_STATES_DONE: &str = "done";
+
 pub static SELECT_BRANCH_POPUP_MSG: &str = "Switch Branch";
 
-pub fn title_status(key_config: &SharedKeyConfig) -> String {
-    format!(
-        "Unstaged Changes [{}]",
-        key_config.get_hint(key_config.focus_workdir)
-    )
+pub fn title_status(_key_config: &SharedKeyConfig) -> String {
+    "Unstaged Changes".to_string()
 }
 pub fn title_diff(_key_config: &SharedKeyConfig) -> String {
     "Diff: ".to_string()
 }
-pub fn title_index(key_config: &SharedKeyConfig) -> String {
-    format!(
-        "Staged Changes [{}]",
-        key_config.get_hint(key_config.focus_stage)
-    )
+pub fn title_index(_key_config: &SharedKeyConfig) -> String {
+    "Staged Changes".to_string()
 }
 pub fn tab_status(key_config: &SharedKeyConfig) -> String {
     format!("Status [{}]", key_config.get_hint(key_config.tab_status))
@@ -101,6 +100,15 @@ pub fn confirm_msg_merge(
 }
 pub fn confirm_msg_reset(_key_config: &SharedKeyConfig) -> String {
     "confirm file reset?".to_string()
+}
+pub fn confirm_msg_reset_lines(
+    _key_config: &SharedKeyConfig,
+    lines: usize,
+) -> String {
+    format!(
+        "are you sure you want to discard {} selected lines?",
+        lines
+    )
 }
 pub fn confirm_msg_stashdrop(
     _key_config: &SharedKeyConfig,
@@ -341,6 +349,16 @@ pub mod commands {
             CMD_GROUP_LOG,
         )
     }
+    pub fn push_tags(key_config: &SharedKeyConfig) -> CommandText {
+        CommandText::new(
+            format!(
+                "Push Tags [{}]",
+                key_config.get_hint(key_config.push),
+            ),
+            "push tags to remote",
+            CMD_GROUP_LOG,
+        )
+    }
     pub fn diff_home_end(
         key_config: &SharedKeyConfig,
     ) -> CommandText {
@@ -373,10 +391,46 @@ pub mod commands {
     ) -> CommandText {
         CommandText::new(
             format!(
-                "Revert hunk [{}]",
+                "Reset hunk [{}]",
                 key_config.get_hint(key_config.status_reset_item),
             ),
             "reverts selected hunk",
+            CMD_GROUP_DIFF,
+        )
+    }
+    pub fn diff_lines_revert(
+        key_config: &SharedKeyConfig,
+    ) -> CommandText {
+        CommandText::new(
+            format!(
+                "Reset lines [{}]",
+                key_config.get_hint(key_config.status_reset_lines),
+            ),
+            "resets selected lines",
+            CMD_GROUP_DIFF,
+        )
+    }
+    pub fn diff_lines_stage(
+        key_config: &SharedKeyConfig,
+    ) -> CommandText {
+        CommandText::new(
+            format!(
+                "Stage lines [{}]",
+                key_config.get_hint(key_config.diff_stage_lines),
+            ),
+            "stage selected lines",
+            CMD_GROUP_DIFF,
+        )
+    }
+    pub fn diff_lines_unstage(
+        key_config: &SharedKeyConfig,
+    ) -> CommandText {
+        CommandText::new(
+            format!(
+                "Unstage lines [{}]",
+                key_config.get_hint(key_config.diff_stage_lines),
+            ),
+            "unstage selected lines",
             CMD_GROUP_DIFF,
         )
     }
@@ -424,18 +478,6 @@ pub mod commands {
         )
         .hide_help()
     }
-    pub fn select_staging(
-        key_config: &SharedKeyConfig,
-    ) -> CommandText {
-        CommandText::new(
-            format!(
-                "To stage [{}]",
-                key_config.get_hint(key_config.focus_stage),
-            ),
-            "focus/select staging area",
-            CMD_GROUP_GENERAL,
-        )
-    }
     pub fn select_status(
         key_config: &SharedKeyConfig,
     ) -> CommandText {
@@ -446,18 +488,6 @@ pub mod commands {
                 key_config.get_hint(key_config.tab_log),
             ),
             "focus/select file tree of staged or unstaged files",
-            CMD_GROUP_GENERAL,
-        )
-    }
-    pub fn select_unstaged(
-        key_config: &SharedKeyConfig,
-    ) -> CommandText {
-        CommandText::new(
-            format!(
-                "To unstaged [{}]",
-                key_config.get_hint(key_config.focus_workdir),
-            ),
-            "focus/select unstaged area",
             CMD_GROUP_GENERAL,
         )
     }